Strict-mode clean: full no-fpermissive build (clang/NDK readiness)
Full 657-target rebuild with GCC without -fpermissive succeeds with zero
errors - a hard requirement for the Android NDK clang build.

Fix catalog applied across ~45 files:
- (int)ptr -> (int)(intptr_t)ptr for EF_Query results, entity id casts,
  CHandle assignment, offsetof-style shader offset probes, sort keys
- rvalue address-of temps hoisted to named locals:
  * DrawTriStrip(&(CVertexBuffer(...))) -> CVertexBuffer vb_(...); 60+ sites
  * pe_status_awake()/pe_action_*() -> named locals passed by pointer
  * &Vec3d(...)/&PropertyWriter(...)/CStreamData_WorldPos hoists
- vectorr (real) alias preserved for triangle-normal abs() probe
- ETexType 0 assignment cast; char* != 0 comparison
- DEFINE_ALIGNED_DATA_STATIC_INIT macro (attribute before '=') for
  initialized aligned matrices (NULL_RendPipeline, GLRendPipeline)
- SDL_GL_GetProcAddress (SDL_FunctionPointer) -> void* via intptr_t
- pipe-water-level pointer diff via byte pointer subtraction

+++ SourceCode/Cry3DEngine/3DEngine.cpp
@@ -459,7 +459,7 @@ void C3DEngine::SetCamera(const CCamera &cam, bool bToTheScreen)
 												*/
 	// under water camera effects
 //  float fUnderWater = GetWaterLevel(&GetViewCamera().GetPos())-GetViewCamera().GetPos().z;
-//  int nRecursionLevel = (int)GetRenderer()->EF_Query(EFQ_RecurseLevel);
+//  int nRecursionLevel = (int)(intptr_t)GetRenderer()->EF_Query(EFQ_RecurseLevel);
 //	float fZMin = GetViewCamera().GetZMin();
   //if(fUnderWater>(-fZMin) && nRecursionLevel==0)
   {
@@ -860,7 +860,7 @@ bool C3DEngine::PhysicalizeStaticObject(void *pForeignData,int iForeignData,int 
 	if (!m_pTerrain || !m_pObjManager)
 		return false;
 
-	int ix=(int)pForeignData&0xFF, iy=(int)pForeignData>>8&0xFF, iobj=(int)pForeignData>>16&0xFFFF;
+	int ix=(int)(intptr_t)pForeignData&0xFF, iy=(int)(intptr_t)pForeignData>>8&0xFF, iobj=(int)(intptr_t)pForeignData>>16&0xFFFF;
 	return 0;//m_pObjManager->PhysicalizeStatObjInst( &m_pTerrain->m_arrSecInfoTable[ix][iy]->m_lstStatObjects[iobj], true );
 }
